fix error

+++ log_transform/kernel.cu
@@ -8,7 +8,7 @@ __global__ void log_transformation_kernel(double* input, double* output, double 
     // INSERT KERNEL CODE HERE
     int i = threadIdx.x + blockDim.x * blockIdx.x;
     printf("Double");
-    printf(log10f(256));
+    printf(unsigned(log10f(256)));
     double c = 255 / log10f(256);
     if (i < size)
     {
